requirements.txt
Dash==2.17.1
plotly==5.22.0
pandas==2.2.2
numpy==1.26.4
Pillow==10.3.0
Flask-Caching==2.3.0
dash-bootstrap-components==1.6.0
dash-canvas==0.1.0
